fix: fix context injection
